// apps/frontend/src/lib/api/generated/schemas/profileProjectSchema.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Resume Genius API
 * OpenAPI spec version: 1.0.0
 */
import type { ProfileProjectSchemaDescription } from "./profileProjectSchemaDescription";
import type { ProfileProjectSchemaStartDate } from "./profileProjectSchemaStartDate";
import type { ProfileProjectSchemaEndDate } from "./profileProjectSchemaEndDate";
import type { ProfileProjectSchemaProjectUrl } from "./profileProjectSchemaProjectUrl";
import type { ProfileProjectSchemaRepositoryUrl } from "./profileProjectSchemaRepositoryUrl";

export interface ProfileProjectSchema {
  id: string;
  user_id: string;
  project_name: string;
  description?: ProfileProjectSchemaDescription;
  start_date?: ProfileProjectSchemaStartDate;
  end_date?: ProfileProjectSchemaEndDate;
  project_url?: ProfileProjectSchemaProjectUrl;
  repository_url?: ProfileProjectSchemaRepositoryUrl;
  created_at: string;
  updated_at: string;
}